feat: emit beforecollapseexpandcombo and aftercollapseexpandcombo; fix: node state and update fontSize problem, closes: #2462
